Adding a score variable
BP_Net has a new variable linked to the score. It helps executing the 3 movements in a certain order. I tested the conditionnal execution by using multiple values.
For now, the net follows this order : a vertical move at the beginning, a horizontal move when the scores reachs a value of 10, and a circular move after gaining 20 points.